requirements.txt
openai
streamlit-chat
streamlit-authenticator
